refactor: use plugin structure for each conc mode
