+++ tox.ini
@@ -5,13 +5,13 @@ skipsdist = True
 
 [testenv]
 commands =
-    pip install ./
+    pip install ./ --upgrade
     py.test {posargs:tests}
 
 [testenv:flake8]
 skip_install = true
 deps = flake8
-commands = flake8 pytest_concurrent.py setup.py tests
+commands = flake8 pytest_concurrent setup.py tests
 
 [flake8]
 ignore = E501
